+++ CHANGELOG.md
@@ -14,6 +14,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 ## [4.10.1]
 ### Added
 ### Changed
+- issue-58: Changed harmony submit call from GET to POST.
 - feature-29: updated docker node version to 20.12.1 (the most current long term support version).
 ### Removed
 ### Fixed